Auto Cleanup

+++ src/main/java/boilerplate/client/fx/FXRaygun.java
@@ -178,14 +178,14 @@ public class FXRaygun extends EntityFX
         tessellator.draw();
         GL11.glPushMatrix();
         float var9 = 1.0F;
-        float slide = (float)this.worldObj.getWorldTime();
-        float rot = (float)(this.worldObj.provider.getWorldTime() % (360 / this.rotationspeed) * this.rotationspeed) + this.rotationspeed * f;
+        float slide = this.worldObj.getWorldTime();
+        float rot = this.worldObj.provider.getWorldTime() % (360 / this.rotationspeed) * this.rotationspeed + this.rotationspeed * f;
         float size = 1.0F;
 
         if (this.pulse)
         {
             size = Math.min(this.particleAge / 4.0F, 1.0F);
-            size = (float)(this.prevSize + (size - this.prevSize) * f);
+            size = this.prevSize + (size - this.prevSize) * f;
         }
 
         float op = 0.4F;
@@ -243,8 +243,8 @@ public class FXRaygun extends EntityFX
         float yy = (float)(prey + (py - prey) * f - EntityFX.interpPosY);
         float zz = (float)(prez + (pz - prez) * f - EntityFX.interpPosZ);
         GL11.glTranslated(xx, yy, zz);
-        float ry = (float)(this.prevYaw + (this.rotYaw - this.prevYaw) * f);
-        float rp = (float)(this.prevPitch + (this.rotPitch - this.prevPitch) * f);
+        float ry = this.prevYaw + (this.rotYaw - this.prevYaw) * f;
+        float rp = this.prevPitch + (this.rotPitch - this.prevPitch) * f;
         GL11.glRotatef(90.0F, 1.0F, 0.0F, 0.0F);
         GL11.glRotatef(180.0F + ry, 0.0F, 0.0F, -1.0F);
         GL11.glRotatef(rp, 1.0F, 0.0F, 0.0F);
