+++ src/main/java/boilerplate/client/fx/FXSmoke.java
@@ -46,14 +46,14 @@ public class FXSmoke extends EntityReddustFX
         this.particleScale *= scale;
         this.smokeParticleScale = this.particleScale;
         this.particleMaxAge = (int)(8.0D / (Math.random() * 0.8D + 0.2D));
-        this.particleMaxAge = (int)((float)this.particleMaxAge * scale);
+        this.particleMaxAge = (int)(this.particleMaxAge * scale);
         this.noClip = false;
     }
 
     @Override
     public void renderParticle(Tessellator tessellator, float par2, float par3, float par4, float par5, float par6, float par7)
     {
-        float f6 = ((float)this.particleAge + par2) / (float)this.particleMaxAge * 32.0F;
+        float f6 = (this.particleAge + par2) / this.particleMaxAge * 32.0F;
 
         if (f6 < 0.0F)
         {
