+++ src/main/java/boilerplate/common/Utils.java
@@ -17,41 +17,41 @@ public class Utils {
 
 		for (int amount = 0; amount < 6; amount++)
 		{
-			double dx = (float)x + random.nextFloat();
-			double dy = (float)y + random.nextFloat();
-			double dz = (float)z + random.nextFloat();
+			double dx = x + random.nextFloat();
+			double dy = y + random.nextFloat();
+			double dz = z + random.nextFloat();
 
 			if ((amount == 0) && (!world.isBlockNormalCubeDefault(x, y + 1, z, false)))
 			{
-				dy = (double)(y + 1) + offset;
+				dy = y + 1 + offset;
 			}
 			
 			if ((amount == 1) && (!world.isBlockNormalCubeDefault(x, y - 1, z, false)))
 			{
-				dy = (double)(y + 0) - offset;
+				dy = y + 0 - offset;
 			}
 			
 			if ((amount == 2) && (!world.isBlockNormalCubeDefault(x, y, z + 1, false)))
 			{
-				dz = (double)(z + 1) + offset;
+				dz = z + 1 + offset;
 			}
 			
 			if ((amount == 3) && (!world.isBlockNormalCubeDefault(x, y, z - 1, false)))
 			{
-				dz = (double)(z + 0) - offset;
+				dz = z + 0 - offset;
 			}
 			
 			if ((amount == 4) && (!world.isBlockNormalCubeDefault(x + 1, y, z, false)))
 			{
-				dx = (double)(x + 1) + offset;
+				dx = x + 1 + offset;
 			}
 			
 			if ((amount == 5) && (!world.isBlockNormalCubeDefault(x - 1, y, z, false)))
 			{
-				dx = (double)(x + 0) - offset;
+				dx = x + 0 - offset;
 			}
 			
-			if ((dx < (double)x) || (dx > (double)(x + 1)) || (dy < 0.0D) || (dy > (double)(y + 1)) || (dz < (double)z) || (dz > (double)(z + 1)))
+			if ((dx < x) || (dx > x + 1) || (dy < 0.0D) || (dy > y + 1) || (dz < z) || (dz > z + 1))
 			{
 				world.spawnParticle(particleName, dx, dy, dz, -1.0D, 1.0D, -1.0D);
 				//Steamcraft.proxy.smokeFX(world, dx, dy, dz, FXSmoke.class);
